+++ test/document/revisions_tests.dart
@@ -39,7 +39,7 @@ class RevisionsTests
 
   RevisionsTests(this.context) {
     remoteDataFolder = context.remoteBaseTestDataFolder + '/DocumentActions/Revisions';
-    localFile = 'Common/test_multi_pages.docx';
+    localFile = 'DocumentElements/Revisions/TestRevisions.doc';
   }
 
   /// Test for accepting revisions in document.
@@ -106,5 +106,35 @@ class RevisionsTests
     expect(result.model, isNotNull);
     expect(result.model?.result, isNotNull);
     expect(result.model?.result?.dest, isNotNull);
+  }
+
+  /// Test for getting revisions from document.
+  Future<void> testGetAllRevisions() async
+  {
+    final remoteFileName = 'TestAcceptAllRevisions.docx';
+    await context.uploadFile(localFile, remoteDataFolder + '/' + remoteFileName);
+
+    final request = GetAllRevisionsRequest(
+      remoteFileName,
+      folder: remoteDataFolder
+    );
+
+    final result = await context.getApi().getAllRevisions(request);
+    expect(result.revisions, isNotNull);
+    expect(result.revisions?.revisions?.length, 6);
+  }
+
+  /// Test for getting revisions online from document.
+  Future<void> testGetAllRevisionsOnline() async
+  {
+    final requestDocument = await context.loadBinaryFile(localFile);
+
+    final request = GetAllRevisionsOnlineRequest(
+      requestDocument
+    );
+
+    final result = await context.getApi().getAllRevisionsOnline(request);
+    expect(result.revisions, isNotNull);
+    expect(result.revisions?.revisions?.length, 6);
   }
 }